Alterações feitas e novo script adicionado

+++ actions/processarExcluirConta.php
@@ -3,25 +3,26 @@
     include_once "../database/database.php";
     //$excluir = filter_input(INPUT_POST, 'btnExcluir', FILTER_SANITIZE_STRING);
     $numeroConta = filter_input(INPUT_POST, 'numeroConta', FILTER_SANITIZE_STRING);
+    $saldo = filter_input(INPUT_POST, 'saldo', FILTER_SANITIZE_STRING);
     if(isset($_POST['btnExcluir'])){
         try{
-            $sql = "DELETE FROM contas WHERE numeroConta = '$numeroConta'";
-            $stmt = $conn->prepare($sql);
-            $stmt->bindValue(":numeroConta", $numeroConta, PDO::PARAM_STR);
-            $stmt->execute();
-            if($conn->query($sql)){
-                header("Location: ../index.php?status=true");
-            }else{
+            if($saldo > 0){
                 header("Location: ../index2.php?status=error");
-                echo "Erro ao deletar conta";
-
             }
-            
+            else{
+                $sql = "DELETE FROM contas WHERE numeroConta = '$numeroConta'";
+                $stmt = $conn->prepare($sql);
+                $stmt->bindValue(":numeroConta", $numeroConta, PDO::PARAM_STR);
+                $stmt->execute();
+                if($conn->query($sql)){
+                    header("Location: ../index.php?status=true");
+                }
+            }
         }catch(PDOException $e){
             echo $e->getMessage();
         }
     }//Xdebug
-echo $numeroConta;
+//echo $numeroConta;
 
     //Deletando com MYSQLI
     /*if($excluir){
